create tasks within your project!
Finally. Spent all day trying to crack it.

send project_id param with task/new link

use before filter in task controller to load up the project_id straight into the new task.

either populate a field in the  form with this, or pass the param with the form and catch it in the create method