Time Logs, Completion
Since introducing validation, timelogs were not saving their comments
this was a combination of;
TL validating presence of comment_id - there was no comment_id because
the foreign key was ime_log_id on the comment table.
I ran a migration to remove the comment_id field

Comments still validate user_id and body, but these were not being set
properly in the TL create action. you cannot mass assign params[] and
then add on extra arguments with comments.

Completion now works by linking straight to the action. - need to make
this a bit prettier. Admins can reverse this with a reopen button